kansei

+++ printf/ft_printf.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libftprintf.h                                      :+:      :+:    :+:   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mrio <mrio@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/05 11:51:46 by mrio              #+#    #+#             */
-/*   Updated: 2025/05/13 15:55:26 by mrio             ###   ########.fr       */
+/*   Updated: 2025/05/13 16:27:52 by mrio             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ size_t	ft_putstr(char *str);
 size_t	ft_putnbr(int c);
 size_t	ft_putnbr_u(unsigned int c);
 size_t	ft_pointer(void *ptr);
-size_t	ft_puthex(unsigned int i, const char str);
+size_t	ft_puthex(unsigned long i, const char str);
 
 #endif
